refactor: subcommand admin avertissement
#145
