+++ slash_commands/subcommands/admin/index.js
@@ -1,5 +1,7 @@
 const { salon } = require("./salon");
+const { avertissement } = require("./avertissement");
 
 module.exports = {
     salon,
+    avertissement,
 };
